Check migration inheritance via ancestors array
In some (perhaps misguided) cases, classes may override the <=> method
(on the class, not the instance). This causes the inheritance check via
the less-than operator to throw an exception.

So instead, let’s use the `ancestors` method, which is less likely to be
altered by other code. A class’s ancestors always includes itself, so we
need to ignore the base class as well.