🌟🗣 grammar glowup of README.md
The name of the project appears to be "Karafka Web", but "Karafka Web UI" (without "the") would only make sense grammatically if that is the name of the project.  I would recommend either adding "The" or else dropping "UI" (as I did). Dropping "UI" makes the following use of "The Web UI" not seem so repetetive as it would if it were immediately following "The Karafka Web UI".  Another alternative would be changing the project name to "Karafka Web UI". 


> All of the Karafka ecosystem components' documentation, including the Web UI, can be found [here](https://karafka.io/docs/#web-ui).

It is possible (though not very likely) that I have misunderstood the intention of the original form of this sentence.  My changes reflect an understanding of it saying that the documentation for all components (plural) within the Karafka ecosystem can be found at the given URL. 


In any grammar "correction" (or suggestion of a change) there is a risk of misunderstanding the original intent.  For example, I would need more information on what is meant by "Real-time topics lag awareness" to be 100% certain that my change is right, but I think you likely want to say "Real-time topic lag awareness".  In English adjectives are not normally pluralized, so my reading of this is that "topic" is the adjective describing "lag", and Karafka gives you monitoring tools to detect topic lag in real time.  Since I'm not a Kafka or Karafka user (yet?)  I don't know what topic lag is (and therefore I don't know with 100% certainty that I'm parsing the grammar correctly). But I'm relatively confident that the change I'm proposing is appropriate.  Of course always feel free to correct me if I'm misunderstanding anything.

"the data produced to Kafka topics" is grammatically unconventional. Do you mean "produced by"?  "produced for"? "sent to"? "produced and sent to"?   I wasn't sure how to change that, so I left it as-is, following the principle to "first do no harm".  

Signed-off-by: Brandon Zylstra <9854+brandonzylstra@users.noreply.github.com>